Rename person directory to student directory
