+++ src/main/java/seedu/address/model/ReadOnlyAddressBook.java
@@ -1,7 +1,7 @@
 package seedu.address.model;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Student;
+import seedu.address.model.student.Student;
 
 /**
  * Unmodifiable view of an address book
