+++ src/main/java/seedu/address/model/student/NameContainsKeywordsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.student;
 
 import java.util.List;
 import java.util.function.Predicate;
