+++ src/main/java/seedu/address/model/student/exceptions/DuplicatePersonException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.student.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Persons (Persons are considered duplicates if they have the same
