+++ src/main/java/seedu/address/model/student/exceptions/PersonNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.student.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified student.
