+++ src/test/java/seedu/address/testutil/TestUtil.java
@@ -7,7 +7,7 @@ import java.nio.file.Paths;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.model.Model;
-import seedu.address.model.person.Student;
+import seedu.address.model.student.Student;
 
 /**
  * A utility class for test cases.
